Convert to ksh and fix $0 for source
Made sure all scripts are ksh for consistency. We have to use ksh
for the associative arrays to work correctly as the code currently
operates.

Also converted $0 to use ${.sh.file} instead. This change is because
if a script is sourced, $0 yields the name of the original script,
not the current script. This syntax is ksh-specific. The equivilant
bash command is ${BASH_SOURCE[0]}.

Refs: #63941, #76266