Fix minor issues with lr segment
A couple of minor bugs crept in to the lr segment during the j-job
revamp. Some of the wave settings for forecast had to be shuffled
between the parm and forecast script so that they are set after
segment-specific changes are made (turning cplwav off). The location
for the stoch_out files for restart was incorrect. And the FHOUR in
post was being set to some unrelated value.

Refs: #63941, #76266