Clean j-job and error messages for gempak and gempak_meta
Remakes the j-job to NCO standards. Places where jid ($$) were used are
replaced with $LSB_JOBID. New parm files are created for each job that
holds some of the settings formerly in the j-job.

Replaced the hard-coded sleep construction with one that mirrors other
scripts.

Additionally, a lot of output and error checking is standardized,
including the PS4 and error messages.

Refs: #63941, #76266