Clean j-job and error messages for wave_gempak
Remakes the j-job to NCO standards. Places where jid ($$) were used are
replaced with $LSB_JOBID. A new script-level script is create to
generate the MPMD job (though it only uses one thread) and some settings
formerly in the j-job were moved to the parm.

Also fixed a bug in the rocoto script where the Dell 3.5 versions were
being loaded on Dell 3.

Additionally, a lot of output and error checking is standardized,
including the PS4 and error messages.

Refs: #63941, #76266