Clean j-job and error messages for ensavg_nemsio
Remakes the j-job to NCO standards. Places where jid ($$) were used are
replaced with $LSB_JOBID. The previous script-level script is pushed
down to the ush level to make room for a new script-level script to hold
some of the code from the former j-job. Some settings are also moved to
a new parm file.

Additionally, a lot of output and error checking is standardized,
including the PS4 and error messages.

Refs: #63941, #76266