Clean j-job and error messages for postsnd
Remakes the j-job to NCO standards. Places where jid ($$) were used are
replaced with $LSB_JOBID. A new parm file is created to hold some of the
settings formerly in the j-job. Others settings are pushed down to the
scripts-level.

Additionally, a lot of output and error checking is standardized,
including the PS4 and error messages.

Refs: #63941, #76266