Fix problems with post_cleanup and add chem version
Some issues with the update of atmos_post_cleanup are fixed.

Also, a new chem version of post_cleanup is added. For this script,
the ush script removing the files in sfcsig is reused, but a new
version of the script-level script is needed because the aerosol
member produces different products than the atmos members.

Refs: #63941, #76266