Fix rocoto resource request for atmos_prep
The previous resource settings for atmos_prep failed when the full
ensemble was run because tasks were not spanning all nodes properly.
Changed the way <nodes> was formulated so that ptile is set
correctly.

Also took this opportunity to determine fewer nodes can be used. As
few as 3 nodes still runs successfully for the full 30+1 ensemble,
but there is a trade-off of runtime. Now using 6 nodes (down from 31),
which seems to be a sweet spot.

Refs: #63941, #76266