Stop creating restart directory on COM if not neeed
The forecast script was creating a restart directory for all forecasts,
even though it is only needed for the aerosol member. Now the restart
directory is only created if cplchm is ".true.". Also placed the
directory and symlink creation in a SENDCOM block so they are not created
unless SENDCOM is YES.

Change-Id: I5efc511a0aa18eaff6ed79e9e76222a2fe168d74
Refs: #63941, #76266