Improve dbn_alert and add a minor fix
Files in init are completely different from v11 and the dbn_alert
commands are added in the new atmos_prep and atmos_init_recdenter
jobs.

Files in bufr are new products and the dbn_alert commands copied
from gfs need modifications. GFS is changed to GEFS and ensemble
member ID is added to the alert message.

Corrections are made to the dbn_alert commands of prdgen/ensstat
ush scripts to match the product and script changes in v12.

A bug is fixed by adding $ to a variable in
cripts/xgefs_atmos_postsnd.sh

Refs: #76266