Fix task geometry for atmos_prep
There were two issues that were causing atmos_prep to not use cpu
efficiently.

First, OMP_NUM_THREADS_CH was not being set correctly in development,
and not at all for operations. The development script now sets the
value correctly depending on how many nodes and members are being used,
and the default operational value for 31 nodes is added to the
atmos_prep parm file.

Second, the affinity directive in the LSF job script needs to use
thread(40) instead of core(1).

A new variable is added to pyGEFS that allows changes to the number of
cores used per task without editing the python code. This is controlled
with the new atmos_prep_cpu_per_task variable, allowing the correct task
geometry to be calculated.

Change-Id: I19d101feb7c2cd271a0919168248a8ad52d1a1e6
Refs: #76266