Fix atmos_prep for GFS v16
The recent merge of develop caused some conflicts due to the
changes made to how surface files are handled there that are
incongruent (and duplicative of) with the way it was being done
in the ncinit branch. Also several things were hardcoded and
needed to be genericized.

The new chgres script (chgres_cube.sh) is a bash script, so the
entire chain is converted to bash to match.

Because chgres_cube is an MPI program, we need to have two
different MPI commands (one for the MPMD, one for the program
inside each MPMD task). At the j-job level, this just means
setting both APRUN and APRUN_MPMD. On the rocoto side, this
means we can't use totaltasks to determine both, so
ATMOS_PREP_CPU_PER_TASK is now exported for the atmos_prep job,
and this is used to set a new gefsmpexec_atmos_prep variable
that is used for APRUN in the atmos_prep job.

chgres_cube also requires the number of MPI tasks is divisible
by the number of tiles (6), so this is checked in the rocoto
system when creating the XML files.

Due to some issue with MPI, when chgres_cube is run on the
hostnode of an MPMD job it never returns, so a dummy command
is used to occupy the first rank. Unfortunately, this requires
the use of an additional node.

As mentioned above, the method for excluding the surface from
the regridding (using DOSFC) in global_chgres_driver is
incompatible with chgres_cube (controlled by CONVERT_SFC). After
the merger, both methods were present; now the DOSFC stuff is
removed. Instead of computing the sfc for all of the perturbed
members and copying the p01 sfc to c00, sfc is only performed on
c00 and copied to the other members.

Change-Id: I512db851fa54cca98b3d6139ced030f12953f219