Revert atmos_prep from MPMD to separate jobs
Since the new chgres job is an MPI job in its own right, and the
running the MPI jobs within an MPMD job was causing computational
inefficiencies for unknown reasons, and was more difficult to
manage in general. The job is not converted back into a metatask
with separate job submissions for each member.

The surface portion of chgres, which was separated out into its own
ush script, is currently added to the control member. A future
commit may merge the surface back in to prevent giving NCO a new
script.

This commit only updates the rocoto system; an update to ecflow to
match will be completed later.

Change-Id: I4a0bb24941601dec077d6095723443aec6dac13b