Update chem init for GFS v16
The chem init job is updated to a new process in order to handle
GFS v16 analysis files. The new process adds an additional step
where the GEFS forecast from the previous cycle is converted from
nemsio to NetCDF. This allows the remaining steps to be performed
using new versions of recenter and calc_increment from GFS v16
that is NetCDF from end-to-end.

The new programs currently require newer (parallel) versions of the
NetCDF and HDF modules. For now, these are hard-coded in the rocoto
bin file to override the versions file, and the job only works on
Dell 3.5. Determination will be made in the future how to reconcile
the version mismatch.

This is also still using a temporary version of global-workflow
from a private repository. Components still need to be tagged
properly and pushed to the authoritative repo.

Change-Id: Ied6665b1cc4fb1c6914ede9702986b39f3e36a04
Refs: #81079